+++ minigu/catalog/src/txn/mod.rs
@@ -2,7 +2,6 @@ pub mod error;
 pub mod manager;
 pub mod transaction;
 pub mod versioned;
-pub mod versioned_map;
 
 pub use error::{CatalogTxnError, CatalogTxnResult};
 pub use manager::CatalogTxnManager;
